* changes:
  Fix return type of QuicSpdyClientBase::latest_response_code
  Change QuicSpdyStreamBodyBuffer API to decouple from QuicStreamSequencer.
  gfe-relnote: Add IsStreamReady() interface to WriteScheduler and implement the interface in subclasses. No functional expected. Not protected.
  Call MarkConsumed() only once in QuicSpdyStreamBodyBuffer::MarkBodyConsumed() with a combined byte count.
  Add debugging logs for Priority frames.

+++ quic/tools/quic_spdy_client_base.h
@@ -120,7 +120,7 @@ class QuicSpdyClientBase : public QuicClientBase,
 
   void set_store_response(bool val) { store_response_ = val; }
 
-  size_t latest_response_code() const;
+  int latest_response_code() const;
   const std::string& latest_response_headers() const;
   const std::string& preliminary_response_headers() const;
   const spdy::SpdyHeaderBlock& latest_response_header_block() const;
